requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
asgiref=3.7.2=pyhd8ed1ab_0
backports.zoneinfo=0.2.1=py310hbe9552e_8
blas=2.119=openblas
blas-devel=3.9.0=19_osxarm64_openblas
blosc=1.21.3=h313beb8_0
boost-cpp=1.74.0=h32e41df_4
bzip2=1.0.8=h620ffc9_4
c-ares=1.21.0=h93a5062_0
ca-certificates=2023.11.17=hf0a4a13_0
cairo=1.16.0=h302bd0f_5
cfitsio=3.470=ha83d327_7
crispy-bootstrap5=0.7=pypi_0
django=4.2.7=pyhd8ed1ab_0
django-crispy-forms=2.1=pypi_0
django-environ=0.11.2=pyhd8ed1ab_0
django-leaflet=0.29.0=pyhd8ed1ab_0
expat=2.5.0=hb7217d7_1
fontconfig=2.14.1=hee714a5_2
freetype=2.10.4=h17b34a0_1
freexl=1.0.6=h1a28f6b_0
gdal=3.6.2=py310h1b5a848_3
geos=3.9.1=h9f76cd9_2
geotiff=1.7.0=h41f0982_1
gettext=0.21.1=h0186832_0
giflib=5.2.1=h1a8c8d9_3
glib=2.69.1=h514c7bf_2
hdf4=4.2.13=h5e329fb_3
hdf5=1.12.1=nompi_had0e5e0_100
icu=68.2=hbdafb3b_0
jpeg=9e=h1a8c8d9_3
json-c=0.16=hc449e50_0
kealib=1.5.0=hba2eb73_1
krb5=1.20.1=h127bd45_0
lcms2=2.15=h481adae_0
lerc=3.0=hc377ac9_0
libblas=3.9.0=19_osxarm64_openblas
libcblas=3.9.0=19_osxarm64_openblas
libcurl=8.2.1=h0f1d93c_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libexpat=2.5.0=hb7217d7_1
libffi=3.4.4=hca03da5_0
libgdal=3.6.2=hb72785c_3
libgfortran=5.0.0=13_2_0_hd922786_1
libgfortran5=13.2.0=hf226fd6_1
libiconv=1.17=he4db4b2_0
libkml=1.3.0=h893ab4d_1014
liblapack=3.9.0=19_osxarm64_openblas
liblapacke=3.9.0=19_osxarm64_openblas
libnetcdf=4.8.1=ha022005_4
libnghttp2=1.52.0=h10c0552_1
libopenblas=0.3.24=openmp_hd76b1f2_0
libpng=1.6.39=h80987f9_0
libpq=12.15=h449679c_1
libspatialite=4.3.0a=h4707d77_23
libssh2=1.10.0=h449679c_2
libtiff=4.5.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=hb547adb_0
libxml2=2.10.4=h372ba2a_0
libzip=1.8.0=h0c481fb_0
llvm-openmp=17.0.4=hcd81f8e_0
lz4-c=1.9.4=hb7217d7_0
ncurses=6.4=h313beb8_0
nspr=4.35=hb7217d7_0
nss=3.89.1=h313beb8_0
numpy=1.26.0=py310h3b2db8e_0
numpy-base=1.26.0=py310ha9811e2_0
openblas=0.3.24=openmp_hce3e5ba_0
openjpeg=2.3.0=h7a6adac_2
openssl=1.1.1w=h53f4e23_0
pcre=8.45=hbdafb3b_0
pcre2=10.42=hb066dcc_0
pillow=10.0.1=py310h3b245a6_0
pip=23.3=py310hca03da5_0
pixman=0.42.2=h13dd4ca_0
poppler=22.12.0=h52f4003_3
poppler-data=0.4.12=hd8ed1ab_0
proj=7.2.0=heac154c_1
psycopg2=2.9.3=py310h80987f9_1
python=3.10.0=hbdb9e5c_5
python_abi=3.10=2_cp310
qhull=2020.2=hc021e02_2
readline=8.2=h1a28f6b_0
setuptools=68.0.0=py310hca03da5_0
sqlite=3.41.2=h80987f9_0
sqlparse=0.4.4=pyhd8ed1ab_0
tiledb=2.3.3=h3ad0f7e_2
tk=8.6.12=hb8d0fd4_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
wheel=0.41.2=py310hca03da5_0
xerces-c=3.2.4=h389e118_0
xz=5.4.2=h80987f9_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
